feat(corgi-web): 身份认证
登录、登出

closes #13
