+++ corgi-web/conf/application.yml.default
@@ -1,0 +1,31 @@
+# Server
+server:
+  display-name: "Corgi 柯基"
+  address: 127.0.0.1
+  port: 8080
+
+# Spring
+spring:
+  datasource:
+    initialize: true
+    initialization-mode: always
+    platform: mysql
+    name: mysql
+    driver-class-name: com.mysql.jdbc.Driver
+    url: jdbc:mysql://127.0.0.1:3306/corgi?useUnicode=true&characterEncoding=utf-8
+    username: test
+    password: test
+  thymeleaf:
+    cache: false
+  cache:
+    ehcache:
+      config: classpath:ehcache.xml
+  devtools:
+    livereload:
+      enabled: true
+
+# Druid
+druid:
+  initial-size: 1
+  min-idle: 1
+  max-active: 20
