+++ corgi-web/conf/cluster.properties.default
@@ -1,0 +1,4 @@
+# Hive
+hive.master.url=jdbc:hive2://127.0.0.1:10000
+hive.master.username=
+hive.master.password=
