+++ corgi-web/pom.xml
@@ -20,6 +20,7 @@
         <druid.version>1.1.9</druid.version>
         <hadoop.version>2.6.5</hadoop.version>
         <hive.version>1.1.0</hive.version>
+        <shiro.version>1.4.0</shiro.version>
 
         <bootstrap.version>4.1.0</bootstrap.version>
         <jquery.version>3.3.1</jquery.version>
@@ -76,7 +77,7 @@
             <optional>true</optional>
         </dependency>
 
-        <!-- Data -->
+        <!-- Database -->
         <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
@@ -103,7 +104,24 @@
             <version>${ehcache.version}</version>
         </dependency>
 
-        <!-- hadoop -->
+        <!-- Shiro -->
+        <dependency>
+            <groupId>org.apache.shiro</groupId>
+            <artifactId>shiro-core</artifactId>
+            <version>${shiro.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.shiro</groupId>
+            <artifactId>shiro-web</artifactId>
+            <version>${shiro.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.shiro</groupId>
+            <artifactId>shiro-spring</artifactId>
+            <version>${shiro.version}</version>
+        </dependency>
+
+        <!-- Hadoop -->
         <dependency>
             <groupId>org.apache.hadoop</groupId>
             <artifactId>hadoop-common</artifactId>
@@ -124,6 +142,10 @@
             <artifactId>hive-jdbc</artifactId>
             <version>${hive.version}</version>
         </dependency>
+
+        <!-- Shiro -->
+
+
 
         <!-- Webjars -->
         <dependency>
