+++ corgi-web/src/main/java/com/dyingbleed/corgi/web/controller/api/DataSourceController.java
@@ -3,7 +3,10 @@ package com.dyingbleed.corgi.web.controller.api;
 import com.dyingbleed.corgi.web.bean.DataSource;
 import com.dyingbleed.corgi.web.service.DataSourceService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 import java.util.Map;
