+++ corgi-web/src/main/java/com/dyingbleed/corgi/web/controller/page/LoginPageController.java
@@ -6,20 +6,20 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 /**
- * 首页
+ * 登录页
  *
- * Created by 李震 on 2018/5/10.
+ * Created by 李震 on 2019/2/2.
  */
 @Controller
-@RequestMapping("/")
-public class HomePageController {
+@RequestMapping("/login")
+public class LoginPageController {
 
     /**
-     * 首页
+     * 登录页
      * */
     @GetMapping
-    public ModelAndView homePage() {
-        ModelAndView modelAndView = new ModelAndView("index");
+    public ModelAndView loginPage() {
+        ModelAndView modelAndView = new ModelAndView("login");
         return modelAndView;
     }
