+++ corgi-web/src/main/java/com/dyingbleed/corgi/web/service/BatchTaskService.java
@@ -1,22 +1,13 @@
 package com.dyingbleed.corgi.web.service;
 
 import com.dyingbleed.corgi.web.bean.BatchTask;
-import com.dyingbleed.corgi.web.mapper.BatchTaskMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 /**
- * 批处理任务
- *
- * Created by 李震 on 2018/5/15.
+ * Created by 李震 on 2019/2/2.
  */
-@Service
-public class BatchTaskService {
-
-    @Autowired
-    private BatchTaskMapper batchTaskMapper;
+public interface BatchTaskService {
 
     /**
      * 新增批量任务
@@ -26,10 +17,7 @@ public class BatchTaskService {
      * @return 批量任务
      *
      * */
-    public BatchTask insertBatchTask(BatchTask task) {
-        this.batchTaskMapper.insertBatchTask(task);
-        return task;
-    }
+    public BatchTask insertBatchTask(BatchTask task);
 
     /**
      * 根据 ID 删除批量任务
@@ -37,9 +25,7 @@ public class BatchTaskService {
      * @param id 批量任务 ID
      *
      * */
-    public void deleteBatchTaskById(Long id) {
-        this.batchTaskMapper.deleteBatchTaskById(id);
-    }
+    public void deleteBatchTaskById(Long id);
 
     /**
      * 修改批量任务
@@ -49,10 +35,7 @@ public class BatchTaskService {
      * @return 批量任务
      *
      * */
-    public BatchTask updateBatchTask(BatchTask task) {
-        this.batchTaskMapper.updateBatchTask(task);
-        return task;
-    }
+    public BatchTask updateBatchTask(BatchTask task);
 
     /**
      * 查询所有批量任务
@@ -60,9 +43,7 @@ public class BatchTaskService {
      * @return 批量任务列表
      *
      * */
-    public List<BatchTask> queryAllBatchTask() {
-        return this.batchTaskMapper.queryAllBatchTask();
-    }
+    public List<BatchTask> queryAllBatchTask();
 
     /**
      * 根据 ID 查询批量任务
@@ -72,9 +53,7 @@ public class BatchTaskService {
      * @return 批量任务
      *
      * */
-    public BatchTask queryBatchTaskById(Long id) {
-        return this.batchTaskMapper.queryBatchTaskById(id);
-    }
+    public BatchTask queryBatchTaskById(Long id);
 
     /**
      * 根据名称查询批量任务
@@ -84,8 +63,6 @@ public class BatchTaskService {
      * @return 批量任务
      *
      * */
-    public BatchTask queryBatchTaskByName(String name) {
-        return this.batchTaskMapper.queryBatchTaskByName(name);
-    }
+    public BatchTask queryBatchTaskByName(String name);
 
 }
