+++ corgi-web/src/main/java/com/dyingbleed/corgi/web/service/HiveService.java
@@ -1,42 +1,13 @@
 package com.dyingbleed.corgi.web.service;
 
 import com.dyingbleed.corgi.web.bean.Column;
-import com.dyingbleed.corgi.web.utils.HiveUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.context.annotation.PropertySource;
-import org.springframework.stereotype.Service;
 
-import java.sql.SQLException;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 
 /**
- * Hive Metadata
- *
- * Created by 李震 on 2018/5/17.
+ * Created by 李震 on 2019/2/2.
  */
-@Service
-@PropertySource("file:${CORGI_HOME}/conf/cluster.properties")
-public class HiveService {
-
-    private static final Logger logger = LoggerFactory.getLogger(HiveService.class);
-
-    @Value("${hive.master.url}")
-    private String hiveMasterUrl;
-
-    @Value("${hive.master.username}")
-    private String hiveMasterUsername;
-
-    @Value("${hive.master.password}")
-    private String hiveMasterPassword;
-
-    @Autowired
-    private BatchTaskService batchTaskService;
+public interface HiveService {
 
     /**
      * 显示 Hive 所有数据库
@@ -44,16 +15,7 @@ public class HiveService {
      * @return 数据库列表
      *
      * */
-    @Cacheable(cacheNames = "sink_db")
-    public List<String> showDBs() {
-        List<String> databases = new LinkedList<>();
-        try {
-            databases.addAll(HiveUtils.showDatabases(this.hiveMasterUrl, this.hiveMasterUsername, this.hiveMasterPassword));
-        } catch (SQLException | ClassNotFoundException e) {
-            logger.error("显示所有 Hive 数据库出错", e);
-        }
-        return databases;
-    }
+    public List<String> showDBs();
 
     /**
      * 显示 Hive 数据库所有表
@@ -63,16 +25,7 @@ public class HiveService {
      * @return 表列表
      *
      * */
-    @Cacheable(cacheNames = "sink_table")
-    public List<String> showTables(String db) {
-        List<String> tables = new LinkedList<>();
-        try {
-            tables.addAll(HiveUtils.showTables(this.hiveMasterUrl, this.hiveMasterUsername, this.hiveMasterPassword, db));
-        } catch (SQLException | ClassNotFoundException e) {
-            logger.error("显示所有 Hive 数据库表出错", e);
-        }
-        return tables;
-    }
+    public List<String> showTables(String db);
 
     /**
      * 显示 Hive 表所有字段
@@ -82,16 +35,6 @@ public class HiveService {
      *
      * @return 表字段
      * */
-    public List<Column> descTable(String db, String table) {
-        List<Column> columns = new LinkedList<>();
-        try {
-            for (Map<String, String> i: HiveUtils.descTable(this.hiveMasterUrl, this.hiveMasterUsername, this.hiveMasterPassword, db, table)) {
-                columns.add(new Column(i.get("name"), i.get("type"), i.get("comment")));
-            }
-        } catch (SQLException | ClassNotFoundException e) {
-            logger.error("显示 Hive 表所有字段出错", e);
-        }
-        return columns;
-    }
+    public List<Column> descTable(String db, String table);
 
 }
