+++ corgi-web/src/main/resources/templates/base.html
@@ -30,6 +30,9 @@
                 <a class="nav-item nav-link" th:classappend="${moduleName == 'datasource'} ?'active'" href="/datasource/">数据源</a>
             </div>
         </div>
+        <div class="navbar-text">
+            <a href="/logout">退出</a>
+        </div>
     </nav>
 </body>
 </html>
